venv compatibility, add requirements.txt
Some newer distributions prevent using the system-provided python with
non-distribution-packaged python libraries. Instead they require use of
a 'venv'. In this case we need to use the venv-provided python binary
and not a hardcoded value. Every contemporary distribution provides
/usr/bin/env as a way to find a binary in the current $PATH instead.

Also, `requirements.txt` is a more standard way of defining the required
libs.